Fixed SYM parsing. Still requires applying.

+++ src/main/java/psyq/sym/ISymObject.java
@@ -2,6 +2,6 @@ package psyq.sym;
 
 public interface ISymObject {
 	long getOffset();
-	byte getTag();
+	int getTag();
 	
 }
