+++ src/main/java/psyq/sym/SymObjectName.java
@@ -3,7 +3,7 @@ package psyq.sym;
 public class SymObjectName extends SymObject {
 	String name;
 	
-	public SymObjectName(long offset, byte tag, String name) {
+	public SymObjectName(long offset, int tag, String name) {
 		super(offset, tag);
 		this.name = name;
 	}
